remove unnecessary casts and tests for null, discovered using NetBeans (#1285)

+++ jme3-android/src/main/java/com/jme3/app/state/VideoRecorderAppState.java
@@ -345,7 +345,7 @@ public class VideoRecorderAppState extends AbstractAppState {
         }
 
         public float getTimePerFrame() {
-            return (float) (1.0f / this.framerate);
+            return 1.0f / this.framerate;
         }
 
         public void update() {
