+++ jme3-blender/src/main/java/com/jme3/scene/plugins/blender/textures/generating/NoiseGenerator.java
@@ -464,7 +464,7 @@ import com.jme3.scene.plugins.blender.textures.generating.TextureGeneratorMusgra
                 sum += t * amp;
             }
 
-            sum *= (float) (1 << noiseDepth) / (float) ((1 << noiseDepth + 1) - 1);
+            sum *= (1 << noiseDepth) / (float) ((1 << noiseDepth + 1) - 1);
             return sum;
         }
