+++ jme3-desktop/src/main/java/com/jme3/app/state/VideoRecorderAppState.java
@@ -330,7 +330,7 @@ public class VideoRecorderAppState extends AbstractAppState {
         }
 
         public float getTimePerFrame() {
-            return (float) (1.0f / this.framerate);
+            return 1.0f / this.framerate;
         }
 
         public void update() {
