+++ jme3-terrain/src/main/java/com/jme3/terrain/heightmap/FluidSimHeightMap.java
@@ -211,7 +211,7 @@ public class FluidSimHeightMap extends AbstractHeightMap {
         // put the normalized heightmap into the range [0...255] and into the heightmap
         for (int y = 0; y < size; y++) {
             for (int x = 0; x < size; x++) {
-                heightData[x + y * size] = (float) (tempBuffer[curBuf][x + y * size]);
+                heightData[x + y * size] = tempBuffer[curBuf][x + y * size];
             }
         }
         normalizeTerrain(NORMALIZE_RANGE);
